+StockEvents & new json format
StockEvents can now be tracked.
The model is able to represent a stockevent. This event denotes a change in stock that is not explained by the ingestion of a medication.

Models now only store the id which is linked to the substance that is referenced. The pharmaceutical that is referenced via the ID will be loaded later. This process is called rehydration.

TODO: awaiting a services/controllers load should wait until all items have been added to the controller