WIPRefactored PharmaceuticalSelector.
The PharmaceuticalSelector was an integral part of AddLogEntry.
But adding a stock item to the stock requires selecting a pharmaceutical too.
Therefore it has been WIP refactored into a Widget.
WIP_NOTE: This needs some more code, testing and refinement.

As for now it handles the searching in the known pharmaceuticals.
TODO:
- handle failed search,
- handle state of the selector.
 - either use a controller or access the state via GlobalKey.

Removed some warnings.