Stockadding functional.
TODO: make the StockState selectable, add the longPress on StockCard to reset callback, Use the stockItemCard in StockView

The optionSelector had the issue that even though the selected callback wasnt called it displayed a selected option. This resulted in differing state between the parent and the widget itself.

The dateTimePicker had the same issue. This was mitigated through the title property. It will be displayed on widgets first build and until the widget has been clicked on and a date / time has been selected.
It has also been expanded to handle selecting time (inclusive OR) selecting date.

Took 3 hours 20 minutes