BackupWriter is WIP
We are currently trying to remove the LogController. It is messy and not single responsibility.
This creates issues like needing to have a complex loader (LogService) and
dos multiple other things.

The only reason this is kind important is bcs it keeps the global nextID.
It should maybe delegate to invisible delegates
that handle almost anything except the ID thingy.

TODO: The Controller (in the future aka Repository)